backend/requirements.txt
annotated-types>=0.6.0
anyio>=3.5.0,<5
bcrypt>=4.0.0
click>=8.0.0
colorama>=0.4.0
fastapi>=0.104.0
google-generativeai>=0.4.0
greenlet>=3.0.0
h11>=0.13.0
idna>=3.0
openai>=1.3.0
mcp>=1.0.0
passlib>=1.7.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
pydantic_core>=2.14.0
SQLAlchemy>=2.0.0
sqlmodel>=0.0.16
starlette>=0.27.0
typing-inspection>=0.4.0
typing_extensions>=4.5.0
uvicorn>=0.24.0
httpx>=0.25.0
python-multipart>=0.0.6
